+++ pages/accounting/_middleware.ts
